+++ eval/lib/jwt.go
@@ -133,6 +133,26 @@ func NewJWTSigningConfigFromJWT(j *config.JWT) (*JWTSigningConfig, error) {
 	return c, nil
 }
 
+var NoOpJwtSignFunction = function.New(&function.Spec{
+	Params: []function.Parameter{
+		{
+			Name: "jwt_signing_profile_label",
+			Type: cty.String,
+		},
+		{
+			Name: "claims",
+			Type: cty.DynamicPseudoType,
+		},
+	},
+	Type: function.StaticReturnType(cty.String),
+	Impl: func(args []cty.Value, _ cty.Type) (ret cty.Value, err error) {
+		if len(args) > 0 {
+			return cty.StringVal(""), fmt.Errorf("missing jwt_signing_profile or jwt (with signing_ttl) block with referenced label %q", args[0].AsString())
+		}
+		return cty.StringVal(""), fmt.Errorf("missing jwt_signing_profile or jwt (with signing_ttl) definitions")
+	},
+})
+
 func NewJwtSignFunction(ctx *hcl.EvalContext, jwtSigningConfigs map[string]*JWTSigningConfig,
 	evalFn func(*hcl.EvalContext, hcl.Expression) (cty.Value, error)) function.Function {
 	return function.New(&function.Spec{
@@ -151,7 +171,7 @@ func NewJwtSignFunction(ctx *hcl.EvalContext, jwtSigningConfigs map[string]*JWTS
 			label := args[0].AsString()
 			signingConfig, exist := jwtSigningConfigs[label]
 			if !exist {
-				return cty.StringVal(""), fmt.Errorf("missing jwt_signing_profile or jwt (with signing_ttl) block with referenced label %q", label)
+				return NoOpJwtSignFunction.Call(args)
 			}
 
 			var claims, argumentClaims, headers map[string]interface{}