+++ handler/proxy_test.go
@@ -2,6 +2,7 @@ package handler_test
 
 import (
 	"context"
+	"net/http"
 	"net/http/httptest"
 	"testing"
 	"time"
@@ -10,6 +11,7 @@ import (
 
 	"github.com/avenga/couper/config"
 	"github.com/avenga/couper/config/body"
+	"github.com/avenga/couper/config/request"
 	"github.com/avenga/couper/eval"
 	"github.com/avenga/couper/handler"
 	"github.com/avenga/couper/handler/transport"
@@ -24,6 +26,7 @@ func TestProxy_BlacklistHeaderRemoval(t *testing.T) {
 			Origin: "https://1.2.3.4/",
 		}, nil, logEntry),
 		&hclsyntax.Body{},
+		false,
 		logEntry,
 	)
 
@@ -31,7 +34,7 @@ func TestProxy_BlacklistHeaderRemoval(t *testing.T) {
 	outreq.Header.Set("Authorization", "Basic 123")
 	outreq.Header.Set("Cookie", "123")
 	outreq = outreq.WithContext(eval.NewContext(nil, &config.Defaults{}, "").WithClientRequest(outreq))
-	ctx, cancel := context.WithDeadline(outreq.Context(), time.Now().Add(time.Millisecond*50))
+	ctx, cancel := context.WithDeadline(context.WithValue(context.Background(), request.RoundTripProxy, true), time.Now().Add(time.Millisecond*50))
 	outreq = outreq.WithContext(ctx)
 	defer cancel()
 
@@ -43,5 +46,57 @@ func TestProxy_BlacklistHeaderRemoval(t *testing.T) {
 
 	if outreq.Header.Get("Cookie") != "" {
 		t.Error("Expected removed Cookie header")
+	}
+}
+
+func TestProxy_WebsocketsAllowed(t *testing.T) {
+	log, _ := test.NewLogger()
+	logEntry := log.WithContext(context.Background())
+
+	origin := test.NewBackend()
+
+	pNotAllowed := handler.NewProxy(
+		transport.NewBackend(body.NewHCLSyntaxBodyWithStringAttr("origin", origin.Addr()), &transport.Config{
+			Origin: origin.Addr(),
+		}, nil, logEntry),
+		&hclsyntax.Body{},
+		false,
+		logEntry,
+	)
+
+	pAllowed := handler.NewProxy(
+		transport.NewBackend(body.NewHCLSyntaxBodyWithStringAttr("origin", origin.Addr()), &transport.Config{
+			Origin: origin.Addr(),
+		}, nil, logEntry),
+		&hclsyntax.Body{},
+		true,
+		logEntry,
+	)
+
+	headers := http.Header{
+		"Connection": []string{"upgrade"},
+		"Upgrade":    []string{"websocket"},
+	}
+
+	outreqN := httptest.NewRequest("GET", "http://couper.local/ws", nil)
+	outreqA := httptest.NewRequest("GET", "http://couper.local/ws", nil)
+
+	outCtx := context.WithValue(context.Background(), request.RoundTripProxy, true)
+
+	for _, r := range []*http.Request{outreqN, outreqA} {
+		for h := range headers {
+			r.Header.Set(h, headers.Get(h))
+		}
+	}
+
+	resN, _ := pNotAllowed.RoundTrip(outreqN.WithContext(outCtx))
+	resA, _ := pAllowed.RoundTrip(outreqA.WithContext(outCtx))
+
+	if resN.StatusCode != http.StatusBadRequest {
+		t.Errorf("expected a bad request on ws endpoint without related headers, got: %d", resN.StatusCode)
+	}
+
+	if resA.StatusCode != http.StatusSwitchingProtocols {
+		t.Errorf("expcted passed Connection and Upgrade header which results in 101, got: %d", resA.StatusCode)
 	}
 }